[test] Don't specify 'ingress' as a keyword
Using 'ingress' keyword seems to hurt performance - it's an ingress type of flow for inbound, true, so it needs to stay on the iface type, but specifying this as a parameter as well may not be advisable. The same may be true for 'egress'. 
There was a good source which mentioned this, can't find it atm but want to test throughput with this removed from the parameters. Superficially, sources with instructions to setup cake don't mention to put it as a parameter, just: the bandwidth; overhead considerations and extras like rtt to aim for and minimum packet size on the line.